Changed a bit of the architecture and started working on the UI

+++ TermPaper_OOP/Interfaces/IPositionable.cs
@@ -8,7 +8,7 @@ namespace TermPaper_OOP.Interfaces
 {
     public interface IPositionable
     {
-        decimal X { get; set; }
-        decimal Y { get; set; }
+        float X { get; set; }
+        float Y { get; set; }
     }
 }
